+++ extensions/system-monitor/src/menubar-system-monitor.tsx
@@ -1,0 +1,152 @@
+import { useRef } from "react";
+import { MenuBarExtra, Icon } from "@raycast/api";
+import { usePromise, runAppleScript } from "@raycast/utils";
+import { useInterval } from "usehooks-ts";
+
+import { cpuUsage as osCpuUsage } from "os-utils";
+import { openActivityMonitorAppleScript } from "./utils";
+import { calculateDiskStorage, getOSInfo } from "./SystemInfo/SystemUtils";
+import { getMemoryUsage } from "./Memory/MemoryUtils";
+import { getNetworkData } from "./Network/NetworkUtils";
+import { getBatteryData } from "./Power/PowerUtils";
+
+import { formatBytes, isObjectEmpty } from "./utils";
+
+export default function Command() {
+  const { data: systemInfo, revalidate: revalidateSystem } = usePromise(async () => {
+    const osInfo = await getOSInfo();
+    const storage = await calculateDiskStorage();
+    return { osInfo, storage };
+  });
+
+  const { data: cupUsage, revalidate: revalidateCpu } = usePromise(() => {
+    return new Promise((resolve) => {
+      osCpuUsage((v) => {
+        resolve(Math.round(v * 100).toString());
+      });
+    });
+  });
+
+  const { data: memoryUsage, revalidate: vevalidateMemory } = usePromise(async () => {
+    const memoryUsage = await getMemoryUsage();
+    const memTotal = memoryUsage.memTotal;
+    const memUsed = memoryUsage.memUsed;
+    const freeMem = memTotal - memUsed;
+
+    return {
+      totalMem: Math.round(memTotal / 1024).toString(),
+      freeMemPercentage: Math.round((freeMem * 100) / memTotal).toString(),
+      freeMem: Math.round(freeMem / 1024).toString(),
+    };
+  });
+
+  const prevProcess = useRef<{ [key: string]: number[] }>({});
+  const { data: networkUsage, revalidate: vevalidateNetwork } = usePromise(async () => {
+    const currProcess = await getNetworkData();
+    let upload = 0;
+    let download = 0;
+
+    if (!isObjectEmpty(prevProcess.current)) {
+      for (const key in currProcess) {
+        let down = currProcess[key][0] - (key in prevProcess.current ? prevProcess.current[key][0] : 0);
+
+        if (down < 0) {
+          down = 0;
+        }
+
+        let up = currProcess[key][1] - (key in prevProcess.current ? prevProcess.current[key][1] : 0);
+
+        if (up < 0) {
+          up = 0;
+        }
+
+        download += down;
+        upload += up;
+      }
+    }
+
+    prevProcess.current = currProcess;
+
+    return {
+      upload,
+      download,
+    };
+  });
+
+  const { data: batteryData, revalidate: vevalidateBattery } = usePromise(async () => {
+    const batteryData = await getBatteryData();
+    const isOnAC = !batteryData.isCharging && batteryData.fullyCharged;
+
+    return {
+      batteryData,
+      isOnAC,
+    };
+  });
+
+  useInterval(() => {
+    revalidateSystem();
+    revalidateCpu();
+    vevalidateMemory();
+    vevalidateNetwork();
+    vevalidateBattery();
+  }, 1000);
+
+  return (
+    <MenuBarExtra icon={{ source: "command-icon.png" }} tooltip="System Monitor">
+      <MenuBarExtra.Section title="System Info">
+        <MenuBarExtra.Item
+          title="macOS"
+          subtitle={`${systemInfo?.osInfo.release}` || "Loading..."}
+          icon={Icon.Finder}
+          onAction={() => {}}
+        />
+      </MenuBarExtra.Section>
+      <MenuBarExtra.Section title="Storage">
+        {systemInfo?.storage.map((disk, index) => (
+          <MenuBarExtra.Item
+            key={index}
+            title={disk.diskName}
+            subtitle={`${disk.totalAvailableStorage} GB available of ${disk.totalSize} GB` || "Loading..."}
+            icon={Icon.HardDrive}
+            onAction={() => runAppleScript(openActivityMonitorAppleScript(4))}
+          />
+        ))}
+      </MenuBarExtra.Section>
+      <MenuBarExtra.Section title="CPU">
+        <MenuBarExtra.Item
+          title="CPU Usage"
+          subtitle={`${cupUsage} %` || "Loading..."}
+          icon={Icon.Monitor}
+          onAction={() => runAppleScript(openActivityMonitorAppleScript(1))}
+        />
+      </MenuBarExtra.Section>
+      <MenuBarExtra.Section title="Memory">
+        <MenuBarExtra.Item
+          title="Memory Usage"
+          subtitle={`${memoryUsage?.freeMemPercentage} % (~ ${memoryUsage?.freeMem} GB)` || "Loading..."}
+          icon={Icon.MemoryChip}
+          onAction={() => runAppleScript(openActivityMonitorAppleScript(2))}
+        />
+      </MenuBarExtra.Section>
+      <MenuBarExtra.Section title="Network">
+        <MenuBarExtra.Item
+          title="Network Usage"
+          subtitle={
+            `↓ ${networkUsage?.download !== undefined ? formatBytes(networkUsage.download) : "0 B"}/s ↑ ${networkUsage?.upload !== undefined ? formatBytes(networkUsage.upload) : "0 B"}/s` ||
+            "Loading..."
+          }
+          icon={Icon.Network}
+          onAction={() => runAppleScript(openActivityMonitorAppleScript(5))}
+        />
+      </MenuBarExtra.Section>
+      <MenuBarExtra.Section title="Power">
+        <MenuBarExtra.Item
+          title="Battery"
+          subtitle={batteryData?.batteryData ? `${batteryData?.batteryData?.batteryLevel} %` : "Loading…"}
+          icon={Icon.Plug}
+          onAction={() => runAppleScript(openActivityMonitorAppleScript(3))}
+        />
+      </MenuBarExtra.Section>
+    </MenuBarExtra>
+  );
+}